chore(react-router): setup react router
[Maintains 172019592]
