+++ src/components/Button.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import classes from '../styles/Button.module.css';
+const button = props =>  <button className={classes.Button}>{props.btnName}</button>
+
+export default button;
