+++ src/styles/Greeting.module.css
@@ -1,0 +1,5 @@
+
+.Heading{
+   margin-top: 30px;
+   margin-bottom: 60px;;
+}
